+++ examples/tests/alarms/multi_alarm_simple_test/main.c
@@ -1,11 +1,12 @@
+#include <inttypes.h>
 #include <stdio.h>
 #include <stdlib.h>
 
 #include <libtock-sync/services/alarm.h>
 
 static void event_cb(uint32_t now, uint32_t expiration, void* ud) {
-  int i = (int)ud;
-  printf("%d %lu %lu\n", i, now, expiration);
+  uintptr_t i = (uintptr_t)ud;
+  printf("%zu %" PRIu32 " %" PRIu32 "\n", i, now, expiration);
 }
 
 int main(void) {
